fix: fix type path
