+++ packages/vchart/src/series/sunburst/interface.ts
@@ -1,7 +1,7 @@
 import type { IAnimationSpec } from '../../animation/spec';
 import type { IMarkSpec, IMarkTheme, ISeriesSpec } from '../../typings/spec';
 import type { SunburstLabelConfig } from '@visactor/vgrammar-hierarchy';
-import type { SunburstAppearPreset, SunburstMark } from './animation';
+import type { SunburstAppearPreset, SunburstMark } from './animation/interface';
 import type { IArcMarkSpec, ITextMarkSpec } from '../../typings';
 import type { IPolarSeriesTheme } from '../polar/interface';
 import type { SeriesMarkNameEnum } from '../interface/type';
